cobro-flow-backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Pydantic v2 and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# File uploads
python-multipart>=0.0.6

# Google Cloud / Vertex AI
google-cloud-aiplatform>=1.74.0
google-cloud-tasks>=2.16.0
google-cloud-storage>=2.18.0

# Document processing
langchain-text-splitters>=0.3.0
tiktoken>=0.8.0
PyPDF2>=3.0.0
python-docx>=1.1.0
numpy>=1.26.0

# Development
python-dotenv>=1.0.0
